# Backend/docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'exam'","description":"'PIII Initial Scaffolding'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'PIII Initial Scaffolding'"}],"paths":{"/api/v1/reservas":{"get":{"tags":["reserva-controller"],"operationId":"getAllReservas","parameters":[{"name":"cliente_id","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"videojuego_id","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"puesto_id","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"fecha","in":"query","required":false,"schema":{"type":"string","format":"date-time"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReservaDTO"}}}}}}},"post":{"tags":["reserva-controller"],"operationId":"createTurno","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservaPostDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservaDTO"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/api/v1/videojuegos":{"get":{"tags":["juego-controlle"],"operationId":"obtenerJuegos","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/videoJuegoDTO"}}}}}}}},"/api/v1/puestos":{"get":{"tags":["puesto-controller"],"operationId":"obtenerPuestos","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PuestoDTO"}}}}}}}},"/api/v1/clientes":{"get":{"tags":["cliente-controller"],"operationId":"obtenerClientes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ClienteDTO"}}}}}}}}},"components":{"schemas":{"ReservaPostDTO":{"type":"object","properties":{"fechaHora":{"type":"string","format":"date-time"},"duracionMinutos":{"type":"integer","format":"int32"},"observaciones":{"type":"string"},"cliente_id":{"type":"integer","format":"int64"},"videojuego_id":{"type":"integer","format":"int64"},"puesto_id":{"type":"integer","format":"int64"}}},"Cliente":{"required":["nombreCompleto"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombreCompleto":{"type":"string"},"email":{"type":"string"}}},"PuestoJuego":{"required":["nombre"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"},"tipo":{"type":"string"}}},"ReservaDTO":{"required":["cliente","duracionMinutos","fechaHora","puesto","videojuego"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"cliente":{"$ref":"#/components/schemas/Cliente"},"videojuego":{"$ref":"#/components/schemas/VideoJuego"},"puesto":{"$ref":"#/components/schemas/PuestoJuego"},"fechaHora":{"type":"string","format":"date-time"},"duracionMinutos":{"type":"integer","format":"int32"},"observaciones":{"type":"string"}}},"VideoJuego":{"required":["titulo"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"titulo":{"type":"string"},"genero":{"type":"string"}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}},"videoJuegoDTO":{"required":["titulo"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"titulo":{"type":"string"},"genero":{"type":"string"}}},"PuestoDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"},"tipo":{"type":"string"}}},"ClienteDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombreCompleto":{"type":"string"},"email":{"type":"string"}}}}}}