# Backend/src/main/java/ar/edu/utn/frc/tup/piii/dtos/common/ClienteDTO.java
package ar.edu.utn.frc.tup.piii.dtos.common;


import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Builder
@Data
@AllArgsConstructor
@NoArgsConstructor
public class ClienteDTO {
    private Long id;
    private String nombreCompleto;
    private String email;
}